requirements.txt
aniso8601==9.0.1
astroid==2.12.13
click==8.1.3
dill==0.3.6
dnspython==2.2.1
Flask==2.2.2
Flask-RESTful==0.3.9
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mongoengine==0.25.0
platformdirs==2.6.2
pylint==2.15.9
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
Werkzeug==2.2.2
wrapt==1.14.1
aniso8601==9.0.1
astroid==2.12.13
click==8.1.3
dill==0.3.6
dnspython==2.2.1
Flask==2.2.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mongoengine==0.25.0
platformdirs==2.6.2
pylint==2.15.9
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.7
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
Werkzeug==2.2.2
wrapt==1.14.1
